--- app/spreadsheet.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-require './app/spreadsheet_table'
-
-class Spreadsheet
-  def initialize
-    @spreadsheet_table = SpreadsheetTable.new
-  end
-
-  def get(cell)
-    @spreadsheet_table.get(cell)
-  end
-
-  def get_literal(cell)
-    @spreadsheet_table.get_literal(cell)
-  end
-
-  def put(cell, value)
-    @spreadsheet_table.put(cell, value)
-  end
-end
